# services/marketplace-api/api/openapi.yaml
openapi: "3.0.3"
info:
  title: Marketplace API
  version: "1.0.0"
  description: REST API для маркетплейса - товары, заказы, авторизация

tags:
  - name: auth
  - name: products
  - name: orders
  - name: promo-codes

paths:
  /auth/register:
    post:
      tags: [auth]
      operationId: Register
      summary: Регистрация пользователя
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'

  /auth/login:
    post:
      tags: [auth]
      operationId: Login
      summary: Вход в систему
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: Успешный вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/Error'
        '401':
          $ref: '#/components/responses/Error'

  /auth/refresh:
    post:
      tags: [auth]
      operationId: RefreshToken
      summary: Обновление токена
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Новая пара токенов
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Error'

  /products:
    get:
      tags: [products]
      operationId: ListProducts
      summary: Список товаров (публичный)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: category
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/ProductStatus'
        - name: seller_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Список товаров
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductListResponse'

    post:
      tags: [products]
      operationId: CreateProduct
      summary: Создать товар (SELLER, ADMIN)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductCreate'
      responses:
        '201':
          description: Товар создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'

  /products/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [products]
      operationId: GetProduct
      summary: Получить товар по ID (публичный)
      responses:
        '200':
          description: Товар
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/Error'

    put:
      tags: [products]
      operationId: UpdateProduct
      summary: Обновить товар (SELLER own, ADMIN)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProductUpdate'
      responses:
        '200':
          description: Обновлённый товар
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'

    delete:
      tags: [products]
      operationId: DeleteProduct
      summary: Удалить товар — архивирование (SELLER own, ADMIN)
      security:
        - BearerAuth: []
      responses:
        '204':
          description: Удалено
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'

  /orders:
    post:
      tags: [orders]
      operationId: CreateOrder
      summary: Создать заказ (USER, ADMIN)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderCreate'
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'

  /orders/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags: [orders]
      operationId: GetOrder
      summary: Получить заказ (USER own, ADMIN)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Заказ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'

    put:
      tags: [orders]
      operationId: UpdateOrder
      summary: Обновить заказ — только в статусе CREATED (USER own, ADMIN)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderUpdate'
      responses:
        '200':
          description: Обновлённый заказ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'
        '429':
          $ref: '#/components/responses/Error'

  /orders/{id}/cancel:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      tags: [orders]
      operationId: CancelOrder
      summary: Отменить заказ (USER own, ADMIN)
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Заказ отменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderResponse'
        '403':
          $ref: '#/components/responses/Error'
        '404':
          $ref: '#/components/responses/Error'
        '409':
          $ref: '#/components/responses/Error'

  /promo-codes:
    post:
      tags: [promo-codes]
      operationId: CreatePromoCode
      summary: Создать промокод (SELLER, ADMIN)
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PromoCodeCreate'
      responses:
        '201':
          description: Промокод создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PromoCodeResponse'
        '400':
          $ref: '#/components/responses/Error'
        '403':
          $ref: '#/components/responses/Error'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    Error:
      description: Ошибка
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'

  schemas:
    ProductStatus:
      type: string
      enum: [ACTIVE, INACTIVE, ARCHIVED]

    OrderStatus:
      type: string
      enum: [CREATED, PAYMENT_PENDING, PAID, SHIPPED, COMPLETED, CANCELED]

    DiscountType:
      type: string
      enum: [PERCENTAGE, FIXED_AMOUNT]

    Role:
      type: string
      enum: [USER, SELLER, ADMIN]

    Error:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties: true

    AuthRegisterRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        password:
          type: string
          minLength: 6
          maxLength: 72
        role:
          $ref: '#/components/schemas/Role'

    AuthLoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string

    AuthResponse:
      type: object
      required: [access_token, refresh_token, token_type]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]

    ProductCreate:
      type: object
      required: [name, price, stock, category]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: double
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          maxLength: 4000
          nullable: true
        price:
          type: number
          format: double
          minimum: 0.01
        stock:
          type: integer
          minimum: 0
        category:
          type: string
          minLength: 1
          maxLength: 100
        status:
          $ref: '#/components/schemas/ProductStatus'

    ProductResponse:
      type: object
      required: [id, name, price, stock, category, status, seller_id, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        price:
          type: number
          format: double
        stock:
          type: integer
        category:
          type: string
        status:
          $ref: '#/components/schemas/ProductStatus'
        seller_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    ProductListResponse:
      type: object
      required: [items, total_elements, page, size]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        total_elements:
          type: integer
        page:
          type: integer
        size:
          type: integer

    OrderItemRequest:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
          maximum: 999

    OrderItemResponse:
      type: object
      required: [id, product_id, quantity, price_at_order]
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
        price_at_order:
          type: number
          format: double

    OrderCreate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
          minItems: 1
          maxItems: 50
        promo_code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
          nullable: true

    OrderUpdate:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemRequest'
          minItems: 1
          maxItems: 50

    OrderResponse:
      type: object
      required: [id, user_id, status, items, total_amount, discount_amount, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/OrderStatus'
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItemResponse'
        total_amount:
          type: number
          format: double
        discount_amount:
          type: number
          format: double
        promo_code_id:
          type: string
          format: uuid
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    PromoCodeCreate:
      type: object
      required: [code, discount_type, discount_value, min_order_amount, max_uses, valid_from, valid_until]
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9_]{4,20}$'
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: double
          minimum: 0.01
        min_order_amount:
          type: number
          format: double
          minimum: 0
        max_uses:
          type: integer
          minimum: 1
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time

    PromoCodeResponse:
      type: object
      required: [id, code, discount_type, discount_value, min_order_amount, max_uses, current_uses, valid_from, valid_until, active]
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        discount_type:
          $ref: '#/components/schemas/DiscountType'
        discount_value:
          type: number
          format: double
        min_order_amount:
          type: number
          format: double
        max_uses:
          type: integer
        current_uses:
          type: integer
        valid_from:
          type: string
          format: date-time
        valid_until:
          type: string
          format: date-time
        active:
          type: boolean